refactor: refactor fetchPost to it's own file
